Cleanup hooks and add Field tests
